separated the rtl code generation from main.rs; cleaned the imports
