+++ module/vs-component/src/sst_sim_gen.rs
@@ -1,4 +1,3 @@
-use core::arch;
 use log::info;
 use minijinja;
 use serde_json;
